requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.4.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
Pillow>=9.5.0
